+++ packages/docs/gatsby-config.js
@@ -1,8 +1,8 @@
 const remarkPlugins = [require('remark-slug')]
 
 module.exports = {
-  __experimentalThemes: ['gatsby-theme-ui'],
   plugins: [
+    'gatsby-plugin-theme-ui',
     'gatsby-plugin-react-helmet',
     'gatsby-plugin-catch-links',
     {
@@ -10,7 +10,6 @@ module.exports = {
       options: {
         extensions: ['.mdx', '.md'],
         remarkPlugins,
-        gatsbyRemarkPlugins: ['gatsby-remark-prismjs'],
       },
     },
   ],