+++ packages/docs/src/gatsby-plugin-theme-ui/components.js
@@ -1,6 +1,8 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui'
 
+import CodeBlock from '@theme-ui/prism'
+
 const heading = Tag => props =>
   !!props.id ? (
     <Tag {...props}>
@@ -12,8 +14,7 @@ const heading = Tag => props =>
           ':hover': {
             textDecoration: 'underline',
           },
-        }}
-      >
+        }}>
         {props.children}
       </a>
     </Tag>
@@ -22,6 +23,8 @@ const heading = Tag => props =>
   )
 
 export default {
+  code: CodeBlock,
+  pre: ({ children }) => <div>{children}</div>,
   h2: heading('h2'),
   h3: heading('h3'),
   h4: heading('h4'),