+++ packages/gatsby-plugin-theme-ui/index.js
@@ -1,1 +1,1 @@
-// noop
+export { default } from './src'