+++ packages/theme-ui/src/index.js
@@ -1,7 +1,7 @@
-export { css, get } from '@styled-system/css'
 export { jsx } from './jsx'
 export { ThemeProvider } from './provider'
 export { Context, useThemeUI } from './context'
 export { ColorMode, useColorMode } from './color-modes'
 export { Styled, components } from './components'
 export { Box, Flex, Layout, Header, Main, Container, Footer } from './layout'
+export { css, get } from '@styled-system/css'